Added referenced libraries to the repository.

+++ Sharpen/Sharpen/Extensions.cs
@@ -605,7 +605,7 @@ namespace Sharpen
 			return p.ExitCode;
 		}
 		
-		public static int GetTotalInFixed (this Inflater inf)
+		public static long GetTotalInFixed (this Inflater inf)
 		{
 			if (inf.TotalIn > 0)
 				return inf.TotalIn + 4;
